refactor scripts and py libraries for el9
