+++ scripts/convertATGCRooStatToJson.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 import json 
 from glob import glob
 import os, sys
